Django 1.8 MySQL
This is sort of lame, but I think we need to check for backticks and
change them to double quotes.  MySQL is alone in having it's own syntax,
backticks `, for quoting names instead of using the sql standard double
quote. There is a MySQL option to change to ansi quotes, however, it
appears that in the django.db MySQL backend, it does generate statements
using backticks, so we will have to check for it.

https://github.com/django/django/blob/master/django/db/backends/mysql/operations.py#L115-L118

Signed-off-by: Steven Githens <swgithen@mtu.edu>